Commenting and refactoring of initial analysis homepage code, added proptypes
